# docs/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/signalement-citoyen/upload-url": {
      "post": {
        "description": "Génère une clé d'objet sous signalements/ et une URL PUT présignée. Le client envoie ensuite le fichier directement à R2, puis fournit la clé retournée en photoKey à POST /signalement-citoyen ou PATCH /signalement-citoyen/:id.",
        "operationId": "SignalementCitoyenController_uploadPhotoUrl",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UploadSignalementPhotoRequestDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "URL présignée générée",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UploadSignalementPhotoResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Non authentifié"
          }
        },
        "security": [
          {
            "JWT": []
          }
        ],
        "summary": "Demander une URL présignée pour la photo d’un signalement",
        "tags": [
          "Signalement Citoyen"
        ]
      }
    },
    "/signalement-citoyen": {
      "post": {
        "description": "Permet à un citoyen authentifié de créer un signalement pour un problème rencontré dans sa ville, à partir d’une clé de photo déjà uploadée sur R2 (optionnelle).",
        "operationId": "SignalementCitoyenController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateSignalementCitoyenDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Le signalement a été créé avec succès",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SignalementCitoyenDto"
                }
              }
            }
          },
          "400": {
            "description": "Données invalides"
          },
          "401": {
            "description": "Non authentifié"
          }
        },
        "security": [
          {
            "JWT": []
          }
        ],
        "summary": "Créer un nouveau signalement citoyen",
        "tags": [
          "Signalement Citoyen"
        ]
      },
      "get": {
        "description": "Retourne une liste paginée de signalements avec possibilité de filtrer par différents critères",
        "operationId": "SignalementCitoyenController_findAll",
        "parameters": [
          {
            "name": "titre",
            "required": false,
            "in": "query",
            "description": "Rechercher par titre du signalement",
            "schema": {
              "example": "Nid de poule",
              "type": "string"
            }
          },
          {
            "name": "search",
            "required": false,
            "in": "query",
            "description": "Recherche plein texte (insensible à la casse) sur le titre, la description, l'adresse et le nom de la catégorie du signalement.",
            "schema": {
              "example": "route abidjan",
              "type": "string"
            }
          },
          {
            "name": "categorieId",
            "required": false,
            "in": "query",
            "description": "Filtrer par identifiant de catégorie",
            "schema": {
              "example": "c1a2t3e4g5o6r7i8e9-0a1b-2c3d-4e5f-6g7h8i9j0k1l",
              "type": "string"
            }
          },
          {
            "name": "statut",
            "required": false,
            "in": "query",
            "description": "Filtrer par statut du signalement",
            "schema": {
              "example": "NOUVEAU",
              "type": "string",
              "enum": [
                "NOUVEAU",
                "EN_COURS",
                "RESOLU",
                "REJETE"
              ]
            }
          },
          {
            "name": "latitude",
            "required": false,
            "in": "query",
            "description": "Filtrer par latitude",
            "schema": {
              "example": 5.36,
              "type": "number"
            }
          },
          {
            "name": "longitude",
            "required": false,
            "in": "query",
            "description": "Filtrer par longitude",
            "schema": {
              "example": -4.0083,
              "type": "number"
            }
          },
          {
            "name": "radiusKm",
            "required": false,
            "in": "query",
            "description": "Rayon de recherche en kilomètres autour de (latitude, longitude). Si fourni avec latitude et longitude, retourne les signalements situés dans ce rayon, triés du plus proche au plus éloigné.",
            "schema": {
              "example": 5,
              "type": "number"
            }
          },
          {
            "name": "citoyenId",
            "required": false,
            "in": "query",
            "description": "Filtrer par identifiant du citoyen",
            "schema": {
              "example": "u1s2e3r4-5i6d-7h8e-9r0e-1a2b3c4d5e6f",
              "type": "string"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "Numéro de la page",
            "schema": {
              "default": 1,
              "example": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Nombre d'éléments par page",
            "schema": {
              "default": 10,
              "example": 10,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Liste des signalements récupérée avec succès",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/PaginatedResponseDto"
                    },
                    {
                      "properties": {
                        "data": {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/SignalementCitoyenDto"
                          }
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "401": {
            "description": "Non authentifié"
          }
        },
        "security": [
          {
            "JWT": []
          }
        ],
        "summary": "Récupérer tous les signalements",
        "tags": [
          "Signalement Citoyen"
        ]
      }
    },
    "/signalement-citoyen/me": {
      "get": {
        "description": "Retourne la liste paginée des signalements créés par l'utilisateur authentifié. L'identifiant du citoyen est déduit du JWT. Utilisé par l'écran profil pour n'afficher que les signalements de l'utilisateur.",
        "operationId": "SignalementCitoyenController_findMine",
        "parameters": [
          {
            "name": "page",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Liste des signalements du citoyen récupérée avec succès",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/PaginatedResponseDto"
                    },
                    {
                      "properties": {
                        "data": {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/SignalementCitoyenDto"
                          }
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "401": {
            "description": "Non authentifié"
          }
        },
        "security": [
          {
            "JWT": []
          }
        ],
        "summary": "Récupérer les signalements de l'utilisateur connecté",
        "tags": [
          "Signalement Citoyen"
        ]
      }
    },
    "/signalement-citoyen/mobile": {
      "get": {
        "description": "Retourne une liste paginée des signalements VALIDÉS (validation = true), destinée à l'application mobile. Supporte les mêmes filtres que GET / (recherche, catégorie, statut, rayon géographique) et la pagination pour le défilement infini.",
        "operationId": "SignalementCitoyenController_findAllMobile",
        "parameters": [
          {
            "name": "titre",
            "required": false,
            "in": "query",
            "description": "Rechercher par titre du signalement",
            "schema": {
              "example": "Nid de poule",
              "type": "string"
            }
          },
          {
            "name": "search",
            "required": false,
            "in": "query",
            "description": "Recherche plein texte (insensible à la casse) sur le titre, la description, l'adresse et le nom de la catégorie du signalement.",
            "schema": {
              "example": "route abidjan",
              "type": "string"
            }
          },
          {
            "name": "categorieId",
            "required": false,
            "in": "query",
            "description": "Filtrer par identifiant de catégorie",
            "schema": {
              "example": "c1a2t3e4g5o6r7i8e9-0a1b-2c3d-4e5f-6g7h8i9j0k1l",
              "type": "string"
            }
          },
          {
            "name": "statut",
            "required": false,
            "in": "query",
            "description": "Filtrer par statut du signalement",
            "schema": {
              "example": "NOUVEAU",
              "type": "string",
              "enum": [
                "NOUVEAU",
                "EN_COURS",
                "RESOLU",
                "REJETE"
              ]
            }
          },
          {
            "name": "latitude",
            "required": false,
            "in": "query",
            "description": "Filtrer par latitude",
            "schema": {
              "example": 5.36,
              "type": "number"
            }
          },
          {
            "name": "longitude",
            "required": false,
            "in": "query",
            "description": "Filtrer par longitude",
            "schema": {
              "example": -4.0083,
              "type": "number"
            }
          },
          {
            "name": "radiusKm",
            "required": false,
            "in": "query",
            "description": "Rayon de recherche en kilomètres autour de (latitude, longitude). Si fourni avec latitude et longitude, retourne les signalements situés dans ce rayon, triés du plus proche au plus éloigné.",
            "schema": {
              "example": 5,
              "type": "number"
            }
          },
          {
            "name": "citoyenId",
            "required": false,
            "in": "query",
            "description": "Filtrer par identifiant du citoyen",
            "schema": {
              "example": "u1s2e3r4-5i6d-7h8e-9r0e-1a2b3c4d5e6f",
              "type": "string"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "Numéro de la page",
            "schema": {
              "default": 1,
              "example": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Nombre d'éléments par page",
            "schema": {
              "default": 10,
              "example": 10,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Liste des signalements validés récupérée avec succès",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/PaginatedResponseDto"
                    },
                    {
                      "properties": {
                        "data": {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/SignalementCitoyenDto"
                          }
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT": []
          }
        ],
        "summary": "Flux mobile : signalements validés uniquement",
        "tags": [
          "Signalement Citoyen"
        ]
      }
    },
    "/signalement-citoyen/{id}": {
      "get": {
        "description": "Retourne les détails complets d'un signalement spécifique",
        "operationId": "SignalementCitoyenController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Identifiant unique du signalement",
            "schema": {
              "example": "a1b2c3d4-e5f6-7g8h-9i0j-k1l2m3n4o5p6",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Signalement trouvé",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SignalementCitoyenDto"
                }
              }
            }
          },
          "401": {
            "description": "Non authentifié"
          },
          "404": {
            "description": "Signalement non trouvé"
          }
        },
        "security": [
          {
            "JWT": []
          }
        ],
        "summary": "Récupérer un signalement par son ID",
        "tags": [
          "Signalement Citoyen"
        ]
      },
      "patch": {
        "description": "Permet de modifier les informations d'un signalement existant, y compris sa photo via une clé R2 déjà uploadée. La validation nécessite les droits administrateur.",
        "operationId": "SignalementCitoyenController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Identifiant unique du signalement à modifier",
            "schema": {
              "example": "a1b2c3d4-e5f6-7g8h-9i0j-k1l2m3n4o5p6",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateSignalementCitoyenDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Signalement mis à jour avec succès",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SignalementCitoyenDto"
                }
              }
            }
          },
          "400": {
            "description": "Données invalides"
          },
          "401": {
            "description": "Non authentifié"
          },
          "403": {
            "description": "Accès réservé aux administrateurs"
          },
          "404": {
            "description": "Signalement non trouvé"
          }
        },
        "security": [
          {
            "JWT": []
          }
        ],
        "summary": "Mettre à jour un signalement (Admin uniquement)",
        "tags": [
          "Signalement Citoyen"
        ]
      },
      "delete": {
        "description": "Supprime définitivement un signalement (soft delete)",
        "operationId": "SignalementCitoyenController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Identifiant unique du signalement à supprimer",
            "schema": {
              "example": "a1b2c3d4-e5f6-7g8h-9i0j-k1l2m3n4o5p6",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Signalement supprimé avec succès",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Signalement supprimé avec succès"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Non authentifié"
          },
          "403": {
            "description": "Accès réservé aux administrateurs"
          },
          "404": {
            "description": "Signalement non trouvé"
          }
        },
        "security": [
          {
            "JWT": []
          }
        ],
        "summary": "Supprimer un signalement (Admin uniquement)",
        "tags": [
          "Signalement Citoyen"
        ]
      }
    },
...
  },
  "info": {
    "title": "Citoyen+ API",
    "description": "API officielle de la plateforme Citoyen+ — application citoyenne de la Côte d'Ivoire. Permet la gestion des actualités, signalements citoyens, bibliothèque de documents et quiz éducatifs.",
    "version": "1.0",
    "contact": {
      "name": "Équipe Citoyen+",
      "url": "https://mec-ci.org",
      "email": "contact@mec-ci.org"
    }
  },
...
  "servers": [
    {
      "url": "https://api.mec-ci.org",
      "description": "Production"
    },
    {
      "url": "http://localhost:8081",
      "description": "Local"
    }
  ],
  "components": {
    "securitySchemes": {
      "JWT": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "in": "header"
      }
    },
    "schemas": {
...
      "PaginatedResponseDto": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "type": "array"
            }
          },
          "meta": {
            "$ref": "#/components/schemas/PaginatedMetaDto"
          }
        },
        "required": [
          "data",
          "meta"
        ]
      },
...
      "UploadSignalementPhotoRequestDto": {
        "type": "object",
        "properties": {
          "filename": {
            "type": "string",
            "description": "Nom du fichier tel qu'il sera envoyé au client (utilisé pour son extension). Le front redimensionne et convertit les images en WebP avant l'upload.",
            "example": "photo.webp"
          },
          "contentType": {
            "type": "string",
            "description": "Type MIME du fichier, figé dans la signature de l’URL présignée",
            "example": "image/webp"
          }
        },
        "required": [
          "filename",
          "contentType"
        ]
      },
      "UploadSignalementPhotoResponseDto": {
        "type": "object",
        "properties": {
          "key": {
            "type": "string",
            "description": "Clé objet R2 générée côté serveur"
          },
          "uploadUrl": {
            "type": "string",
            "description": "URL PUT présignée vers laquelle envoyer le fichier"
          },
          "expiresIn": {
            "type": "number",
            "description": "Durée de validité de l'URL présignée, en secondes"
          }
        },
        "required": [
          "key",
          "uploadUrl",
          "expiresIn"
        ]
      },
      "CreateSignalementCitoyenDto": {
        "type": "object",
        "properties": {
          "titre": {
            "type": "string",
            "description": "Titre du signalement",
            "example": "Nid de poule sur la route principale"
          },
          "description": {
            "type": "string",
            "description": "Description détaillée du signalement",
            "example": "Un grand nid de poule situé au milieu de la chaussée, dangereux pour les véhicules"
          },
          "categorieId": {
            "type": "string",
            "description": "Identifiant de la catégorie du signalement",
            "example": "11000000-0000-0000-0000-000000000001"
          },
          "adresse": {
            "type": "string",
            "description": "Adresse du lieu du signalement",
            "example": "Avenue 12, Abidjan, Côte d'Ivoire"
          },
          "latitude": {
            "type": "number",
            "description": "Latitude de la localisation",
            "example": 5.36
          },
          "longitude": {
            "type": "number",
            "description": "Longitude de la localisation",
            "example": -4.0083
          },
          "citoyenId": {
            "type": "string",
            "description": "Identifiant du citoyen ayant créé le signalement",
            "example": "u1s2e3r4-5i6d-7h8e-9r0e-1a2b3c4d5e6f"
          },
          "photoKey": {
            "type": "string",
            "description": "Clé R2 de la photo, obtenue via une demande d'URL présignée préalable (POST /signalement-citoyen/upload-url)",
            "example": "signalements/1732000000000.jpg"
          }
        },
        "required": [
          "titre",
          "description",
          "categorieId",
          "adresse",
          "latitude",
          "longitude"
        ]
      },
...
      "SignalementCitoyenDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Identifiant unique du signalement",
            "example": "a1b2c3d4-e5f6-7g8h-9i0j-k1l2m3n4o5p6"
          },
          "titre": {
            "type": "string",
            "description": "Titre du signalement",
            "example": "Nid de poule sur la route principale"
          },
          "description": {
            "type": "string",
            "description": "Description détaillée du signalement",
            "example": "Un grand nid de poule situé au milieu de la chaussée, dangereux pour les véhicules"
          },
          "categorieId": {
            "type": "string",
            "description": "Identifiant de la catégorie du signalement",
            "example": "11000000-0000-0000-0000-000000000001"
          },
          "categorie": {
            "description": "Catégorie du signalement",
            "allOf": [
              {
                "$ref": "#/components/schemas/CategorieSignalementDto"
              }
            ]
          },
          "validation": {
            "type": "boolean",
            "description": "Indique si le signalement a été validé",
            "example": false,
            "default": false
          },
          "adresse": {
            "type": "string",
            "description": "Adresse du lieu du signalement",
            "example": "Avenue 12, Abidjan, Côte d'Ivoire"
          },
          "latitude": {
            "type": "number",
            "description": "Latitude de la localisation",
            "example": 5.36
          },
          "longitude": {
            "type": "number",
            "description": "Longitude de la localisation",
            "example": -4.0083
          },
          "photo": {
            "type": "string",
            "description": "URL de la photo du signalement",
            "example": "https://cdn.mec-ci.org/signalements/1732000000000.jpg"
          },
          "statut": {
            "type": "string",
            "description": "Statut actuel du signalement",
            "enum": [
              "NOUVEAU",
              "EN_COURS",
              "RESOLU",
              "REJETE"
            ],
            "example": "NOUVEAU"
          },
          "citoyenId": {
            "type": "string",
            "description": "Identifiant du citoyen ayant créé le signalement",
            "example": "u1s2e3r4-5i6d-7h8e-9r0e-1a2b3c4d5e6f"
          },
          "citoyen": {
            "type": "object",
            "description": "Informations du citoyen ayant créé le signalement"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string",
            "description": "Date de création du signalement",
            "example": "2026-01-29T10:00:00.000Z"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string",
            "description": "Date de dernière mise à jour",
            "example": "2026-01-29T10:00:00.000Z"
          },
          "deletedAt": {
            "format": "date-time",
            "type": "string",
            "description": "Date de suppression (soft delete)",
            "example": null
          }
        },
        "required": [
          "id",
          "titre",
          "description",
          "categorieId",
          "validation",
          "adresse",
          "latitude",
          "longitude",
          "statut",
          "createdAt",
          "updatedAt"
        ]
      },
...
    }
  }
}